Add a 'fileCharSource', and some examples that use file IO.
The 'fileCharSource' function takes a file path and returns a
source of characters. It opens the file and attempts to close it
when the end of file is reached. To be really safe, it should
catch exceptions and try to close the file before re-throwing them,
but I'll leave that for now.

I've also changed 'fileCharSource' so it does not attempt to
close the file handle. This is consistent with John Millikin's
enumerators, and with the notion that Handles should be created
and closed by code at the same level.

I've also simplified the types a little so that we assume IO is
the monad, instead of allowing any monad in the MonadIO class.

Finally, I've added some examples in the 'pure1' executable, which
really ought to be renamed. I'll leave that for later too.